+++ core/runtime/ops.rs
@@ -13,12 +13,23 @@ use std::ptr::NonNull;
 use v8::WriteOptions;
 
 use super::op_driver::OpDriver;
+use super::op_driver::OpScheduling;
 use super::op_driver::V8RetValMapper;
 
 /// The default string buffer size on the stack that prevents mallocs in some
 /// string functions. Keep in mind that Windows only offers 1MB stacks by default,
 /// so this is a limited resource!
 pub const STRING_STACK_BUFFER_SIZE: usize = 1024 * 8;
+
+fn op_scheduing(lazy: bool, deferred: bool) -> OpScheduling {
+  if lazy {
+    OpScheduling::Lazy
+  } else if deferred {
+    OpScheduling::Deferred
+  } else {
+    OpScheduling::Eager
+  }
+}
 
 #[inline(always)]
 pub fn map_async_op_infallible<R: 'static>(
@@ -30,31 +41,14 @@ pub fn map_async_op_infallible<R: 'static>(
   rv_map: V8RetValMapper<R>,
 ) -> Option<R> {
   let pending_ops = &ctx.context_state().pending_ops;
-  if lazy {
-    pending_ops.submit_op_infallible::<_, true, false>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      promise_id,
-      op,
-      rv_map,
-    )
-  } else if deferred {
-    pending_ops.submit_op_infallible::<_, false, true>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      promise_id,
-      op,
-      rv_map,
-    )
-  } else {
-    pending_ops.submit_op_infallible::<_, false, false>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      promise_id,
-      op,
-      rv_map,
-    )
-  }
+  pending_ops.submit_op_infallible_scheduling(
+    op_scheduing(lazy, deferred),
+    ctx.id,
+    ctx.metrics_enabled(),
+    promise_id,
+    op,
+    rv_map,
+  )
 }
 
 #[inline(always)]
@@ -67,34 +61,15 @@ pub fn map_async_op_fallible<R: 'static, E: Into<Error> + 'static>(
   rv_map: V8RetValMapper<R>,
 ) -> Option<Result<R, E>> {
   let pending_ops = &ctx.context_state().pending_ops;
-  if lazy {
-    pending_ops.submit_op_fallible::<_, _, true, false>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      ctx.get_error_class_fn,
-      promise_id,
-      op,
-      rv_map,
-    )
-  } else if deferred {
-    pending_ops.submit_op_fallible::<_, _, false, true>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      ctx.get_error_class_fn,
-      promise_id,
-      op,
-      rv_map,
-    )
-  } else {
-    pending_ops.submit_op_fallible::<_, _, false, false>(
-      ctx.id,
-      ctx.metrics_enabled(),
-      ctx.get_error_class_fn,
-      promise_id,
-      op,
-      rv_map,
-    )
-  }
+  pending_ops.submit_op_fallible_scheduling(
+    op_scheduing(lazy, deferred),
+    ctx.id,
+    ctx.metrics_enabled(),
+    ctx.get_error_class_fn,
+    promise_id,
+    op,
+    rv_map,
+  )
 }
 
 macro_rules! try_number_some {
